+++ dev/formatter.nix
@@ -1,0 +1,10 @@
+{ ... }:
+{
+  perSystem =
+    { pkgs
+    , ...
+    }:
+    {
+      formatter = pkgs.treefmt;
+    };
+}