+++ dev/templates.nix
@@ -1,0 +1,7 @@
+{ inputs, ... }:
+{
+  # nix flake init -t github:terranix/terranix#flake
+  flake.templates = inputs.terranix-examples.templates // {
+    default = inputs.terranix-examples.defaultTemplate;
+  };
+}